refactor: wrap existing stories into dedicated folders
- Reorganized story files into individual folders within `src/stories`
- Ensured all stories and documentation files are correctly nested in their respective directories

BREAKING CHANGE: Story paths have been updated; ensure imports and references are adjusted accordingly.
