+++ src/stories/Button/Button.tsx
@@ -30,12 +30,7 @@ interface ButtonProps {
 export function Button({ primary = false, size = 'medium', backgroundColor, label, ...props }: ButtonProps) {
   const mode = primary ? s.primary : s.secondary;
   return (
-    <button
-      type="button"
-      className={[s.button, s[size], mode].join(' ')}
-      style={{ backgroundColor }}
-      {...props}
-    >
+    <button type="button" className={[s.button, s[size], mode].join(' ')} style={{ backgroundColor }} {...props}>
       {label}
     </button>
   );
