+++ src/stories/Page/Page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Header } from './Header';
+import { Header } from '../Header/Header';
 import './page.css';
 
 type User = {
